Review base system

+++ .termux/optimizer.sh
@@ -5,10 +5,15 @@
 dumpsys deviceidle disable
 dumpsys battery reset
 
-## Set CPU governors to performance
-for gov_file in /sys/devices/system/cpu/cpu[0-9]*/cpufreq/scaling_governor; do
-  echo performance >"$gov_file" 2>/dev/null
-done
+## Set CPU governor to performance for all CPUs
+(
+  for cpu in /sys/devices/system/cpu/cpu[0-9]*; do
+    gov_file="$cpu/cpufreq/scaling_governor"
+    if [ -f "$gov_file" ]; then
+      echo performance >"$gov_file" 2>/dev/null
+    fi
+  done
+) &
 
 ## Set I/O scheduler to noop
 for blockdev in /sys/block/*/queue/scheduler; do
@@ -21,13 +26,19 @@ echo 80 >/proc/sys/vm/swappiness 2>/dev/null
 ## Set oom_score_adj for Android apps to prevent them from being killed
 (
   while true; do
-    for uid in $(ls /proc | grep '^[0-9]' | xargs -I{} sh -c "cat /proc/{}/status 2>/dev/null | grep '^Uid:' | awk '{print \$2}'" | sort -u); do
-      if id "$uid" 2>/dev/null | grep -q 'u0_a'; then
-        for pid in $(pgrep -u "$uid"); do
-          echo -1000 >/proc/$pid/oom_score_adj 2>/dev/null
-        done
+    TERMUX_PID=$(pidof com.termux)
+    if [ -n "$TERMUX_PID" ]; then
+      echo -1000 >/proc/$TERMUX_PID/oom_score_adj 2>/dev/null
+    fi
+
+    for pid in /proc/[0-9]*; do
+      uid=$(grep '^Uid:' "$pid/status" 2>/dev/null | awk '{print $2}')
+      if [ -n "$uid" ]; then
+        username=$(id "$uid" 2>/dev/null)
+        echo "$username" | grep -q 'u0_a' && echo -1000 >"$pid/oom_score_adj" 2>/dev/null
       fi
     done
+
     sleep 32
   done
 ) &
